Cast unsigned int type in printing to pass with gcc-5

+++ Source/CNTKv2LibraryDll/Common.cpp
@@ -253,7 +253,7 @@ namespace CNTK
                 return{ tempCPUDataView->WritableDataBuffer<ElementType>(), tempCPUDataView };
             }
             
-            LogicError("Invalid device type (%u).", deviceType);
+            LogicError("Invalid device type (%u).", (unsigned int)deviceType);
         }
 
         template <typename ElementType>
